+++ features/step_definitions/iva_steps/government_steps.rb
@@ -3,19 +3,19 @@ Given(/^I should be on the government verification screen$/) do
   find(@iva_govt.header_css).visible?
 end
 
-Given(/^I submit the government verification form$/) do
-  @iva_govt.verify(email_type: "unique")
-  @iva_govt.click_verify_button
-end
-
 Given(/^I submit the empty government form$/) do
   @iva_govt = IVAGovernment.new
   @iva_govt.verify(populate: false)
 end
 
-Given(/^I submit the government verification form as a dupe record$/) do
+Given(/^I submit the government verification form as a "([^"]*)" record$/) do |type|
   @iva_govt = IVAGovernment.new
-  @iva_govt.verify(email_type: "dupe")
+  @iva_govt.verify(email_type: type)
+
+  if type == "unique"
+    @iva_govt.click_verify_button
+  end
+
 end
 
 Given(/^I clear the current Government session$/) do
@@ -24,8 +24,4 @@ Given(/^I clear the current Government session$/) do
   browser.manage.delete_all_cookies
   browser.navigate.refresh
   step 'I visit IDP through the "government" policy'
-end
-
-Given(/^I submit the government verification form as a denied record$/) do
-  IVAGovernment.new.verify(email_type: "denied")
 end