+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 All notable changes to this project will be documented in this file. This project does not adhere to [Semantic Versioning](http://semver.org/)
 ## 2.0.8 (06/08/2016)
 - QA-221: Add Seaworld Waves of Honor Partner Integration Test
+- HOTFIX: Refactored some Govt test code
 
 ## 2.0.7 (06/07/2016)
 - HOTFIX: Fixed Police doc upload test