+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Change Log
 
 All notable changes to this project will be documented in this file. This project does not adhere to [Semantic Versioning](http://semver.org/)
+## 2.0.9 (06/10/2016)
+- QA-259: Military doc denied test
+
 ## 2.0.8 (06/08/2016)
 - QA-221: Add Seaworld Waves of Honor Partner Integration Test
 - HOTFIX: Refactored some Govt test code